requirements.txt
matplotlib==3.4.1
streamlit==0.80.0
feyn==1.5.3
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
scikit_learn==0.24.1
